refactor: extract ShadowedReservedNameWarning and move parameter-definition helpers to common
Share the shadowed-reserved warning as a single component, and relocate
getExploreParameterDefinitions/getReferencedParameterDefinitions to
@lightdash/common alongside getAvailableParametersFromTables so the
autocomplete hook no longer imports across the component boundary.
